LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.3_Future_Reserved_Words/S7.5.3_A1.9-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Use of reserved word 'enum'
S7.5.3_A1.9

PASS Expected parsing failure

TEST COMPLETE
